+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - `only: []` and `except: []` options that allow to controll what attributes are rendered/parsed
 - `render_nil: true` option that allows to render nil values
+- Allow to pass a context object to extractor/generator methods
 
 ### Changed
 - Pass whole document to methods for JSON/YAML/TOML so its behavior is consistent with XML mapping